Adds rollable attribute support
Adds the ability to define rollable attributes in the ISDL.

This change introduces a new "roll" parameter to the AttributeExp, allowing
users to specify an expression that is evaluated when the attribute is rolled.
A dice icon is displayed when hovering over the attribute,
and clicking it triggers the roll and displays the result in the chat.

Also includes minor fixes to skill lookups in weapons.